refactor: move Carry entity files to new directory structure
